[issues-9] Реализация модального окна "Предупреждение"

+++ src/app/pages/articles/article/article.component.ts
@@ -5,6 +5,9 @@ import { ArticlesHttpService } from 'src/app/pages/articles/services/articles-ht
 import { IDifficultyOptions, ITagOptions } from 'src/app/shared/models/article.interfaces';
 import { IArticle, IArticleInfo } from 'src/app/shared/models/article.interfaces';
 import { ListsHttpService } from 'src/app/shared/services/lists-http.service';
+import { MatDialog } from '@angular/material/dialog';
+import { WarningComponent } from 'src/app/shared/components/modals/warning/warning.component';
+import { IModalData } from 'src/app/shared/models/modal.interfaces';
 import { TagsService } from 'src/app/shared/services/tags.service';
 
 @Component({
@@ -24,6 +27,7 @@ export class ArticleComponent implements OnInit, OnDestroy {
     private route: ActivatedRoute,
     private articlesHttp: ArticlesHttpService,
     private listsHttp: ListsHttpService,
+    private dialog: MatDialog,
     private tagsService: TagsService,
   ) {
     this.articleId = this.route.snapshot.params['id'] ?? '';
@@ -98,10 +102,17 @@ export class ArticleComponent implements OnInit, OnDestroy {
     this.isArticleView = true;
   }
 
-  // TODO: добавить модалку с предупреждением, когда будет готова
   deleteArticle() {
+    const dialog = this.dialog.open<WarningComponent, IModalData, boolean>(WarningComponent, {
+      data: {
+        content: 'Вы уверены, что хотите удалить статью?',
+      }
+    });
     this.sub.add(
-      this.articlesHttp.deleteArticle(this.articleId)
+      dialog.afterClosed()
+      .subscribe((isDelete) => {
+        if (isDelete) { this.articlesHttp.deleteArticle(this.articleId); }
+      })
     );
   }
 }
